// test/integration/test/ports/prisma/functional/relation-mode-gh-1-to-n/_fixture/cascade-plain/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly PostOneToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'];
    };
    readonly PostOptionalOneToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly UserOneToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly enabled: CodecTypes['pg/bool@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly PostOneToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'];
    };
    readonly PostOptionalOneToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly UserOneToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly enabled: CodecTypes['pg/bool@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly postOneToMany: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
    };
    readonly postOptionalOneToMany: {
      readonly authorId: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
    };
    readonly userOneToMany: {
      readonly enabled: CodecTypes['pg/bool@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly postOneToMany: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
    };
    readonly postOptionalOneToMany: {
      readonly authorId: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
    };
    readonly userOneToMany: {
      readonly enabled: CodecTypes['pg/bool@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;